+++ Makefile.am
@@ -391,7 +391,8 @@ libuv_la_SOURCES += src/unix/android-ifaddrs.c \
                     src/unix/procfs-exepath.c \
                     src/unix/pthread-fixes.c \
                     src/unix/random-getrandom.c \
-                    src/unix/random-sysctl-linux.c
+                    src/unix/random-sysctl-linux.c \
+                    src/unix/epoll.c
 endif
 
 if CYGWIN
@@ -472,7 +473,8 @@ libuv_la_SOURCES += src/unix/linux-core.c \
                     src/unix/procfs-exepath.c \
                     src/unix/proctitle.c \
                     src/unix/random-getrandom.c \
-                    src/unix/random-sysctl-linux.c
+                    src/unix/random-sysctl-linux.c \
+                    src/unix/epoll.c
 test_run_tests_LDFLAGS += -lutil
 endif
